move symbolic Ed25519
