+++ symbolic-base/symbolic-base.cabal
@@ -196,7 +196,7 @@ library
       ZkFold.Symbolic.Data.Combinators
       ZkFold.Symbolic.Data.Conditional
       ZkFold.Symbolic.Data.DiscreteField
-      ZkFold.Symbolic.Data.Ed25519
+      ZkFold.Symbolic.Data.EllipticCurve.Ed25519
       ZkFold.Symbolic.Data.EllipticCurve.Secp256k1
       ZkFold.Symbolic.Data.Eq
       ZkFold.Symbolic.Data.FFA
