+++ symbolic-examples/src/Examples/Commitment.hs
@@ -4,7 +4,7 @@ module Examples.Commitment (
 
 import           ZkFold.Base.Data.Vector           (Vector)
 import           ZkFold.Base.Protocol.IVC.Commit   (HomomorphicCommit (..))
-import           ZkFold.Symbolic.Data.Ed25519      (Ed25519_Point)
+import           ZkFold.Symbolic.Data.EllipticCurve.Ed25519      (Ed25519_Point)
 import           ZkFold.Symbolic.Data.FieldElement (FieldElement)
 
 exampleCommitment
