+++ symbolic-examples/src/Examples/Ed25519.hs
@@ -6,7 +6,7 @@ import           ZkFold.Base.Algebra.Basic.Class
 import           ZkFold.Base.Algebra.EllipticCurve.Ed25519 (Ed25519_Base)
 import           ZkFold.Symbolic.Class                     (Symbolic)
 import           ZkFold.Symbolic.Data.Combinators          (RegisterSize (Auto))
-import           ZkFold.Symbolic.Data.Ed25519
+import           ZkFold.Symbolic.Data.EllipticCurve.Ed25519
 import           ZkFold.Symbolic.Data.FFA                  (KnownFFA)
 import           ZkFold.Symbolic.Data.FieldElement
